refactor: import 경로를 업데이트하여 CartItem 및 상수 파일을 정리
